+++ src/com/intellij/idea/plugin/hybris/tools/logging/actions/LoggerActions.kt
@@ -19,19 +19,19 @@
 package com.intellij.idea.plugin.hybris.tools.logging.actions
 
 import com.intellij.idea.plugin.hybris.common.HybrisConstants
+import com.intellij.idea.plugin.hybris.common.utils.HybrisIcons
 import com.intellij.idea.plugin.hybris.notifications.Notifications
+import com.intellij.idea.plugin.hybris.tools.logging.CxLoggerAccess
 import com.intellij.idea.plugin.hybris.tools.logging.LogLevel
 import com.intellij.idea.plugin.hybris.tools.remote.RemoteConnectionService
 import com.intellij.idea.plugin.hybris.tools.remote.RemoteConnectionType
-import com.intellij.idea.plugin.hybris.tools.remote.execution.logging.LoggingExecutionClient
 import com.intellij.idea.plugin.hybris.tools.remote.execution.logging.LoggingExecutionContext
-import com.intellij.idea.plugin.hybris.util.PackageUtils
 import com.intellij.openapi.actionSystem.ActionUpdateThread
 import com.intellij.openapi.actionSystem.AnAction
 import com.intellij.openapi.actionSystem.AnActionEvent
 import com.intellij.openapi.components.service
 
-abstract class AbstractLoggerAction(private val logLevel: LogLevel) : AnAction(logLevel.name, null, logLevel.icon) {
+abstract class AbstractLoggerAction(private val logLevel: LogLevel) : AnAction(logLevel.name, "", logLevel.icon) {
 
     override fun getActionUpdateThread() = ActionUpdateThread.BGT
 
@@ -53,37 +53,15 @@ abstract class AbstractLoggerAction(private val logLevel: LogLevel) : AnAction(l
             logLevel = logLevel
         )
 
-        project.service<LoggingExecutionClient>().execute(context) { coroutineScope, result ->
-            val abbreviationLogIdentifier = PackageUtils.abbreviatePackageName(logIdentifier)
-
-            if (result.statusCode == 200) {
-                Notifications.info(
-                    "Log level updated",
-                    """
-                        <p>Level  : $logLevel</p>
-                        <p>Logger : $abbreviationLogIdentifier</p>
-                        <p>${server.shortenConnectionName()}</p>
-                        """.trimIndent()
-                )
-                    .hideAfter(5)
-                    .notify(project)
-            } else {
-                Notifications.error(
-                    "Failed to update log level",
-                    """
-                        <p>Level  : $logLevel</p>
-                        <p>Logger : $abbreviationLogIdentifier</p>
-                        <p>${server.shortenConnectionName()}</p>
-                        """.trimIndent()
-                )
-            }
-        }
+        CxLoggerAccess.getInstance(project).setLogger(logIdentifier, logLevel)
     }
 
     override fun update(e: AnActionEvent) {
         super.update(e)
         val isRightPlace = "GoToAction" != e.place
-        e.presentation.isEnabled = isRightPlace
+        val project = e.project ?: return
+
+        e.presentation.isEnabled = isRightPlace && project.service<CxLoggerAccess>().canRefresh
         e.presentation.isVisible = isRightPlace
     }
 
@@ -98,3 +76,21 @@ class WarnLoggerAction : AbstractLoggerAction(LogLevel.WARN)
 class ErrorLoggerAction : AbstractLoggerAction(LogLevel.ERROR)
 class FatalLoggerAction : AbstractLoggerAction(LogLevel.FATAL)
 class SevereLoggerAction : AbstractLoggerAction(LogLevel.SEVERE)
+
+class FetchLoggerStateAction : AnAction("Fetch Logger State", "", HybrisIcons.Log.Action.FETCH) {
+
+    override fun actionPerformed(e: AnActionEvent) {
+        val project = e.project ?: return
+        val loggerAccessService = project.service<CxLoggerAccess>()
+
+        loggerAccessService.fetch()
+    }
+
+    override fun update(e: AnActionEvent) {
+        val isRightPlace = "GoToAction" != e.place
+        val project = e.project ?: return
+
+        e.presentation.isEnabled = isRightPlace && project.service<CxLoggerAccess>().canRefresh
+        e.presentation.isVisible = isRightPlace
+    }
+}